Refactor careers
